Rearranged classes.
