+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/dsl/pipeops/ShallowPipeOperations.scala
@@ -1,7 +1,7 @@
 package io.shiftleft.queryprimitives.dsl.pipeops
 
-import io.shiftleft.queryprimitives.dsl.RealPipe.RealPipe
-import io.shiftleft.queryprimitives.dsl.ShallowPipe.ShallowPipe
+import io.shiftleft.queryprimitives.dsl.pipetypes.RealPipe.RealPipe
+import io.shiftleft.queryprimitives.dsl.pipetypes.ShallowPipe.ShallowPipe
 
 import scala.collection.GenTraversableOnce
