+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/dsl/pipetypes/ShallowPipe.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.queryprimitives.dsl
+package io.shiftleft.queryprimitives.dsl.pipetypes
 
 object ShallowPipe {
   type ShallowPipe[+ElemType] = ElemType
